@@ -2,21 +2,17 @@
   "name": "dbd-companion",
   "version": "0.1.0",
   "private": true,
-  "engines": {
-    "npm": "6.14.11",
-    "node": "14.15.5"
-  },
   "dependencies": {
-    "@apollo/client": "^3.3.20",
-    "bootstrap": "^4.6.0",
-    "graphql": "^15.5.1",
-    "http-proxy-middleware": "^2.0.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.0",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "5.0.1",
-    "sass": "^1.55.0",
+    "@apollo/client": "^3.7.8",
+    "bootstrap": "^4.6.2",
+    "graphql": "^15.8.0",
+    "http-proxy-middleware": "^2.0.6",
+    "react": "^18.2.0",
+    "react-bootstrap": "^1.6.6",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.8.1",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.58.1",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
@@ -44,13 +40,14 @@
     ]
   },
   "devDependencies": {
-    "@types/babel__core": "^7.1.19",
+    "@testing-library/react": "^14.0.0",
+    "@types/babel__core": "^7.20.0",
     "@types/babel__generator": "^7.6.4",
-    "@types/jest": "^26.0.23",
-    "@types/node": "^12.20.13",
-    "@types/react": "^17.0.8",
-    "@types/react-dom": "^17.0.5",
-    "@types/react-router-dom": "^5.1.7",
-    "typescript": "^4.3.2"
+    "@types/jest": "^26.0.24",
+    "@types/node": "^12.20.55",
+    "@types/react": "^17.0.53",
+    "@types/react-dom": "^17.0.19",
+    "@types/react-router-dom": "^5.3.3",
+    "typescript": "^4.9.5"
   }
 }
